+++ NEWS.md
@@ -3,6 +3,7 @@
 - Use dpkg options --force-confdef and --force-confold when installing
   packages using the apt provider
 - Add OpenBSD package provider
+- Add OpenBSD user provider
 
 ## 0.5.0 (2015-09-21)
 - Add attribute hash support to DSL and resources